Removed suggestions to use accurate percentages in ini file
Removed the suggestions to use accurate percentages in the ini file as for this game (all UE4 games on PS4?) the UE4 patch method has the same restrictions as hex editing. (NUL characters can be used in the ini method, e.g. for 75% 100 becomes 75NUL, not sure if it's relevant to mention here)